requirements.txt
cffi==1.15.0
evdev==1.5.0
MouseInfo==0.1.3
mss==6.1.0
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-xlib==0.31
python3-xlib==0.15
pytweening==1.0.4
six==1.16.0
sounddevice==0.4.4
tk==0.1.0
